v2 dictionary search moved to separate file We keep it separate since it contains OS dependent code.
